+++ stroom-app/src/main/java/stroom/app/App.java
@@ -26,19 +26,17 @@ import stroom.config.app.AppConfig;
 import stroom.config.app.Config;
 import stroom.config.app.StroomYamlUtil;
 import stroom.config.global.impl.ConfigMapper;
-import stroom.dashboard.impl.ActiveQueriesWebSocket;
 import stroom.dropwizard.common.Filters;
 import stroom.dropwizard.common.HealthChecks;
 import stroom.dropwizard.common.ManagedServices;
 import stroom.dropwizard.common.RestResources;
 import stroom.dropwizard.common.Servlets;
 import stroom.dropwizard.common.SessionListeners;
+import stroom.dropwizard.common.WebSockets;
 import stroom.event.logging.rs.api.RestResourceAutoLogger;
 import stroom.util.config.AppConfigValidator;
 import stroom.util.config.ConfigValidator;
 import stroom.util.config.PropertyPathDecorator;
-import stroom.util.date.DateUtil;
-import stroom.util.guice.GuiceUtil;
 import stroom.util.io.DirProvidersModule;
 import stroom.util.io.FileUtil;
 import stroom.util.io.HomeDirProvider;
@@ -62,7 +60,6 @@ import io.dropwizard.jersey.sessions.SessionFactoryProvider;
 import io.dropwizard.servlets.tasks.LogConfigurationTask;
 import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
-import io.dropwizard.websockets.WebsocketBundle;
 import org.eclipse.jetty.server.session.SessionHandler;
 import org.eclipse.jetty.servlets.CrossOriginFilter;
 import org.glassfish.jersey.logging.LoggingFeature;
@@ -81,7 +78,6 @@ import javax.servlet.FilterRegistration;
 import javax.servlet.SessionCookieConfig;
 import javax.sql.DataSource;
 import javax.validation.ValidatorFactory;
-import javax.validation.constraints.NotNull;
 
 public class App extends Application<Config> {
 
@@ -97,6 +93,8 @@ public class App extends Application<Config> {
     private Filters filters;
     @Inject
     private Servlets servlets;
+    @Inject
+    private WebSockets webSockets;
     @Inject
     private SessionListeners sessionListeners;
     @Inject
@@ -149,9 +147,6 @@ public class App extends Application<Config> {
         // This allows us to use env var templating and relative (to stroom home) paths in the YAML configuration.
         bootstrap.setConfigurationSourceProvider(StroomYamlUtil.createConfigurationSourceProvider(
                 bootstrap.getConfigurationSourceProvider(), true));
-
-        // Add websockets.
-        bootstrap.addBundle(new WebsocketBundle(ActiveQueriesWebSocket.class));
 
         // Add the GWT UI assets.
         bootstrap.addBundle(new DynamicAssetsBundle(
@@ -267,6 +262,9 @@ public class App extends Application<Config> {
         // Add servlets
         servlets.register();
 
+        // Add web sockets
+        webSockets.register();
+
         // Add session listeners.
         sessionListeners.register();
 
@@ -327,7 +325,7 @@ public class App extends Application<Config> {
         final AppConfigValidator appConfigValidator = validationOnlyInjector.getInstance(AppConfigValidator.class);
 
         LOGGER.info("Validating application configuration file {}",
-                configFile.toAbsolutePath().normalize().toString());
+                configFile.toAbsolutePath().normalize());
 
         final ConfigValidator.Result<AbstractConfig> result = appConfigValidator.validateRecursively(appConfig);