+++ stroom-dashboard/stroom-dashboard-impl/src/main/java/stroom/dashboard/impl/ActiveQueriesWebSocket.java
@@ -1,8 +1,11 @@
 package stroom.dashboard.impl;
 
 import stroom.query.api.v2.QueryKey;
+import stroom.security.api.SecurityContext;
 import stroom.util.logging.LambdaLogger;
 import stroom.util.logging.LambdaLoggerFactory;
+import stroom.util.shared.EntityServiceException;
+import stroom.util.shared.IsWebSocket;
 
 import com.codahale.metrics.annotation.ExceptionMetered;
 import com.codahale.metrics.annotation.Metered;
@@ -16,11 +19,8 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
+import javax.inject.Inject;
 import javax.websocket.CloseReason;
-import javax.websocket.OnClose;
-import javax.websocket.OnError;
-import javax.websocket.OnMessage;
-import javax.websocket.OnOpen;
 import javax.websocket.Session;
 import javax.websocket.server.ServerEndpoint;
 
@@ -28,17 +28,23 @@ import javax.websocket.server.ServerEndpoint;
 @Timed
 @ExceptionMetered
 @ServerEndpoint(ActiveQueriesWebSocket.PATH)
-public class ActiveQueriesWebSocket {
+public class ActiveQueriesWebSocket extends AuthenticatedWebSocket implements IsWebSocket {
 
     private static final LambdaLogger LOGGER = LambdaLoggerFactory.getLogger(ActiveQueriesWebSocket.class);
 
     public static final String PATH = "/active-queries-ws";
 
-    private static ActiveQueriesManager activeQueriesManager;
-
+    private final SecurityContext securityContext;
+    private final ActiveQueriesManager activeQueriesManager;
     private final Map<String, Set<String>> activeQueries = new ConcurrentHashMap<>();
 
-    public ActiveQueriesWebSocket() {
+    @Inject
+    public ActiveQueriesWebSocket(final SecurityContext securityContext,
+                                  final ActiveQueriesManager activeQueriesManager) {
+        super(securityContext);
+        this.securityContext = securityContext;
+        this.activeQueriesManager = activeQueriesManager;
+
         Executors.newScheduledThreadPool(1).schedule(() -> {
             if (activeQueriesManager != null) {
                 activeQueries.values().forEach(set -> {
@@ -52,21 +58,27 @@ public class ActiveQueriesWebSocket {
         }, 1, TimeUnit.MINUTES);
     }
 
-    @OnOpen
     public void onOpen(final Session session) throws IOException {
         LOGGER.debug(() -> "Opening web socket at " + PATH
                 + ", sessionId: " + session.getId()
                 + ", maxIdleTimeout: " + session.getMaxIdleTimeout());
+
+        // Ensure a user is logged in.
+        checkLogin();
+
         // Keep alive forever.
         session.setMaxIdleTimeout(0);
         session.getAsyncRemote().sendText("welcome");
     }
 
-    @OnMessage
     public void onMessage(final Session session, final String message) {
         LOGGER.debug(() -> "Received message on web socket at " + PATH
                 + ", sessionId: " + session.getId()
                 + ", message: [" + message + "]");
+
+        // Ensure a user is logged in.
+        checkLogin();
+
         session.getAsyncRemote().sendText(message.toUpperCase());
 
         if (message.startsWith("add:")) {
@@ -86,12 +98,10 @@ public class ActiveQueriesWebSocket {
         }
     }
 
-    @OnError
     public void onError(final Session session, final Throwable thr) {
         LOGGER.error(thr.getMessage(), thr);
     }
 
-    @OnClose
     public void onClose(final Session session, final CloseReason cr) {
         LOGGER.debug(() -> "Closing web socket at " + PATH
                 + ", sessionId: " + session.getId());
@@ -101,7 +111,14 @@ public class ActiveQueriesWebSocket {
         }
     }
 
-    public static void setActiveQueriesManager(final ActiveQueriesManager aqm) {
-        activeQueriesManager = aqm;
+    private void checkLogin() {
+        try {
+            if (!securityContext.isLoggedIn()) {
+                throw new EntityServiceException("No user is logged in");
+            }
+        } catch (final RuntimeException e) {
+            LOGGER.error(e.getMessage(), e);
+            throw e;
+        }
     }
 }